requirements.txt
asyncpg==0.23.0
aiohttp==3.8.1
SQLAlchemy==1.4.31
fastapi==0.70.0
uvicorn==0.15.0
pydantic==1.9.0
alembic==1.9.1
pytest==7.1.1
asgi-lifespan==1.0.1
httpx==0.22.0
pytest-asyncio==0.18.3
factory-boy==3.2.1
Faker==13.3.4
h11==0.12